requirements.txt
# ── PDF Parsing ───────────────────────────────────────────────────────────────
PyMuPDF==1.24.5

# ── LlamaIndex Core + Integrations ───────────────────────────────────────────
llama-index-core==0.10.68
llama-index-embeddings-huggingface==0.3.1
llama-index-llms-ollama==0.3.4
llama-index-vector-stores-chroma==0.2.0

# ── Vector Store ──────────────────────────────────────────────────────────────
chromadb==0.5.23

# ── Local Embedding Backend ───────────────────────────────────────────────────
sentence-transformers==3.1.1
torch==2.4.1

# ── Data Validation ───────────────────────────────────────────────────────────
pydantic==2.8.2

# ── Sample PDF Generation (PoC utility only) ──────────────────────────────────
fpdf2==2.7.9
